CLI for stage 1 and relationships for stage 2

+++ src/main/java/org/example/repositories/TrainingSessionRepository.java
@@ -8,21 +8,19 @@ import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 
 public class TrainingSessionRepository {
+
     public void saveTrainingSession(TrainingSession trainingSession) {
-
-
         SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
         Session session = sessionFactory.openSession();
         session.beginTransaction();
         Transaction transaction = session.getTransaction();
 
-
         session.save(trainingSession);
 
         transaction.commit();
         session.close();
-
     }
+
     public void updateTrainingSession(TrainingSession trainingSession) {
         SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
         Session session = sessionFactory.openSession();
@@ -31,38 +29,27 @@ public class TrainingSessionRepository {
         session.update(trainingSession);
         session.getTransaction().commit();
         session.close();
-
     }
+
     public void deleteTrainingSession(TrainingSession trainingSession) {
         SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
         Session session = sessionFactory.openSession();
         session.beginTransaction();
-        session.delete(trainingSession);
+        if (trainingSession != null) {
+            session.delete(trainingSession);
+        }
         session.getTransaction().commit();
-        session.close();//
-
+        session.close();
     }
-    public Trainer findById(int id) {
+
+    public TrainingSession findById(int id) {
         SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
         Session session = sessionFactory.openSession();
         session.beginTransaction();
-
-        Trainer trainer = session.get(Trainer.class, id);
+        TrainingSession trainingSession = session.get(TrainingSession.class, id);
         session.getTransaction().commit();
         session.close();
-        return trainer;
-    }
-    public void deleteTrainingSession(int id) {
-        SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
-        Session session = sessionFactory.openSession();//1
-        session.beginTransaction();
-        Trainer trainer = findById(id);
-        if (trainer != null) {
-            session.delete(trainer);
-        }
-
-        session.getTransaction().commit();
-        session.close();
+        return trainingSession;
     }
 
 }
